Removes ckan_proxy and handles some edge cases in tests
