--- src/ckan_proxy/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class CkanProxyConfig(AppConfig):
-    name = 'ckan_proxy'
