+++ src/datasets/tests/test_create.py
@@ -1,9 +1,10 @@
+import uuid
+
 from django.test import TestCase
 from django.contrib.auth import get_user_model
 from django.urls import reverse
 from datasets.models import Dataset, Organisation
 
-from ckan_proxy.util import test_user_key
 
 
 class DatasetsTestCase(TestCase):
@@ -12,7 +13,7 @@ class DatasetsTestCase(TestCase):
         self.test_user = get_user_model().objects.create(
             email="test-signin@localhost",
             username="Test User Signin",
-            apikey=test_user_key()
+            apikey=str(uuid.uuid4())
         )
         self.test_user.set_password("password")
         self.test_user.save()
