+++ src/datasets/tests/test_organisation.py
@@ -1,7 +1,8 @@
+import uuid
+
 from django.test import TestCase
 from django.contrib.auth import get_user_model
 from django.urls import reverse
-from ckan_proxy.util import test_user_key
 
 from datasets.models import Organisation
 from datasets.logic import organisations_for_user
@@ -12,7 +13,7 @@ class OrganisationTestCase(TestCase):
         self.test_user = get_user_model().objects.create(
             email="test-signin@localhost",
             username="Test User Signin",
-            apikey=test_user_key()
+            apikey=str(uuid.uuid4())
         )
         self.test_user.set_password("password")
         self.test_user.save()
