+++ src/datasets/util.py
@@ -5,7 +5,6 @@ from mimetypes import guess_extension
 from django.utils.text import slugify
 
 from datasets.models import Dataset
-from ckan_proxy.logic import dataset_show
 
 import requests
 
@@ -61,8 +60,7 @@ def calculate_dates_for_year(year):
 
 
 def convert_to_slug(title):
-    """ Checks for a local dataset with this name, and if not
-    found also checks the CKAN install to make sure it isn't in use """
+    """ Checks for a local dataset with this name """
     t = title
 
     slug = None
@@ -78,10 +76,7 @@ def convert_to_slug(title):
         try:
             draft = Dataset.objects.get(name=slug)
         except Dataset.DoesNotExist:
-            dataset = dataset_show(slug)
-
-        if not draft and not dataset:
-            break
+            return slug
 
         t = "{}{}".format(title, counter)
         counter += 1
