+++ src/datasets/views.py
@@ -10,13 +10,8 @@ from django.http import HttpResponseRedirect, Http404
 
 
 import datasets.forms as f
-from .logic import organisations_for_user
+from datasets.logic import organisations_for_user
 from datasets.models import Dataset, Datafile
-from ckan_proxy.convert import draft_to_ckan, ckan_to_draft
-from ckan_proxy.logic import (organization_show,
-                              dataset_show,
-                              dataset_create,
-                              dataset_update)
 
 
 def new_dataset(request):
@@ -246,13 +241,7 @@ def edit_addfile_annually(request, dataset_name):
 
 
 def edit_files(request, dataset_name):
-    try:
-        dataset = Dataset.objects.get(name=dataset_name)
-    except Dataset.DoesNotExist:
-        # Try and load the dataset from the live CKAN
-        dataset = ckan_to_draft(dataset_name)
-
-
+    dataset = get_object_or_404(Dataset, name=dataset_name)
 
     url = _frequency_redirect_to(dataset)
 
@@ -286,11 +275,7 @@ def edit_notifications(request, dataset_name):
 
 
 def check_dataset(request, dataset_name):
-    try:
-        dataset = Dataset.objects.get(name=dataset_name)
-    except Dataset.DoesNotExist:
-        # Try and load the dataset from the live CKAN
-        dataset = ckan_to_draft(dataset_name)
+    dataset = get_object_or_404(Dataset, name=dataset_name)
 
     organisation = dataset.organisation
     organisations = organisations_for_user(request.user)
@@ -299,7 +284,8 @@ def check_dataset(request, dataset_name):
 
     if request.method == 'POST':
         dataset.published = True
-        published_date = datetime.now()
+        dataset.published_date = datetime.now()
+        dataset.save()
 
         return HttpResponseRedirect('/manage?newset=1')
