+++ src/userauth/tests/test_signout.py
@@ -1,8 +1,9 @@
+import uuid
+
 from django.test import TestCase
 from django.contrib.auth import get_user_model
 from django.urls import reverse
 
-from ckan_proxy.util import test_user_key
 
 
 class SignoutTestCase(TestCase):
@@ -11,7 +12,7 @@ class SignoutTestCase(TestCase):
         self.test_user = get_user_model().objects.create(
             email="test-signout@localhost",
             username="Test User Signin",
-            apikey=test_user_key()
+            apikey=str(uuid.uuid4())
         )
         self.test_user.set_password("password")
         self.test_user.save()
