+++ src/userauth/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render, redirect
 from django.contrib.auth import authenticate, login, logout
 
-from ckan_proxy.logic import organization_list_for_user
+from datasets.logic import organisations_for_user
 from userauth.logic import set_orgs_for_user
 from .forms import SigninForm
 
@@ -17,11 +17,6 @@ def login_view(request):
             user = authenticate(username=email, password=password)
             if user is not None:
                 login(request, user)
-
-                refined_orgs = [(org['name'], org['title'],)
-                                for org in organization_list_for_user(user)]
-                set_orgs_for_user(request, refined_orgs)
-
                 return redirect("/")
             else:
                 login_failed = True
